+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## [Unreleased]
 
-
+- Drop py3.7 compat.
+- Swap out the deprecated appdirs for platformdirs.